Now we use double to store int part of a parsing number in Json::loadNumber()
We need it in case we are parsing double with big integer part (bigger then std::numeric_limits<int>::max())

+++ src/json.cpp
@@ -41,7 +41,8 @@ Node loadNumber(istream& input)
         input.ignore(1);
     }
 
-    int intPart = 0;
+    // use double to store int part in case we are parsing double with int part bigger then INT_MAX
+    double intPart = 0;
     while (isdigit(input.peek()))
     {
         intPart *= 10;
